refactor: [#661] move UDP Tracker Client mod
