+++ src/console/clients/mod.rs
@@ -1,3 +1,4 @@
 //! Console clients.
 pub mod checker;
 pub mod http;
+pub mod udp;
